chanhe key to be long long

+++ inc/dg_cantor_pairing.h
@@ -1,6 +1,6 @@
 #ifndef DG_CANTOR_PAIRING_H
 #define DG_CANTOR_PAIRING_H
 
-int Get_key_fun(int i, int j, int k);
+long long int Get_key_fun(int i, int j, int k);
 
 #endif
